+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Unreleased
 
+- feat(remix): Add Remix wizard (#387)
+
+Set up the Sentry Remix SDK in your app with one command:
+
+```sh
+npx @sentry/wizard -i remix
+```
+
 - fix(cordova): Fallback to the default Sentry CLI path if not defined. (#401)
 
 ## 3.9.2